server/requirements.txt
asgiref==3.2.3
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
nltk==3.4.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
